created Cell and Square component classes
